+++ apps/web/src/hooks/useAutoSlippage.tsx
@@ -121,16 +121,13 @@ const applySlippageLimits = (
   max = MAX_SLIPPAGE_NUMERATOR,
 ) => {
   if (calculatedSlippage.greaterThan(new Percent(max, 10_000))) {
-    console.log('Auto Slippage: Using MAX_AUTO_SLIPPAGE_TOLERANCE', new Percent(max, 10_000).toFixed(2))
     return new Percent(max, 10_000)
   }
 
   if (calculatedSlippage.lessThan(new Percent(min, 10_000))) {
-    console.log('Auto Slippage: Using MIN_AUTO_SLIPPAGE_TOLERANCE', new Percent(min, 10_000).toFixed(2))
     return new Percent(min, 10_000)
   }
 
-  console.log('Auto Slippage: Using calculated result', calculatedSlippage.toFixed(2))
   return calculatedSlippage
 }
 
@@ -183,7 +180,6 @@ export default function useClassicAutoSlippageTolerance(trade?: SupportedTrade):
     ],
     queryFn: () => {
       if (!trade || onL2) {
-        console.log('Auto Slippage: Using DEFAULT_AUTO_SLIPPAGE because', !trade ? 'no trade' : 'on L2')
         return DEFAULT_AUTO_SLIPPAGE
       }
 
@@ -203,18 +199,6 @@ export default function useClassicAutoSlippageTolerance(trade?: SupportedTrade):
           outputBaseSlippage.lessThan(inputBasedSlippage)
 
         const finalSlippage = shouldUseInputBase ? inputBasedSlippage : outputBaseSlippage
-
-        console.info('Auto Slippage: Calculated result', {
-          shouldUseInputBase,
-          inputBasedSlippage,
-          outputBaseSlippage,
-          dollarCostToUse,
-          inputDollarValue,
-          outputDollarValue,
-          fraction: dollarCostToUse / outputDollarValue,
-          resultBasisPoints: Math.floor((dollarCostToUse / outputDollarValue) * 10000),
-          result: finalSlippage.toFixed(2),
-        })
 
         return finalSlippage
       }