+++ apps/web/src/quoter/quoter.types.ts
@@ -88,7 +88,8 @@ export type QuoteQuery = Options & {
   nonce?: number
   placeholderHash?: string
   for?: string
-  createTime?: number
+  createTime: number
+  routeKey?: string
 }
 
 export interface StrategyQuery {