+++ apps/web/src/views/SwapSimplify/InfinitySwap/FormMainV4ForHomePage.tsx
@@ -159,6 +159,8 @@ function useDefaults(): { inputCurrencyId: string | undefined; outputCurrencyId:
         field: parsed.independentField,
         inputCurrencyId: parsed[Field.INPUT].currencyId,
         outputCurrencyId: parsed[Field.OUTPUT].currencyId,
+        inputChainId: parsed[Field.INPUT].chainId,
+        outputChainId: parsed[Field.OUTPUT].chainId,
         recipient: null,
       }),
     )