+++ apps/web/src/views/SwapSimplify/V4Swap/CrossChainConfirmSwapModal/components/EstimatedTime/index.tsx
@@ -1,0 +1,15 @@
+import dayjs from 'dayjs'
+import { useMemo } from 'react'
+
+export function EstimatedTime({ expectedFillTimeSec }: { expectedFillTimeSec?: number }) {
+  const estimatedTimeDisplay = useMemo(() => {
+    if (!expectedFillTimeSec) return '-'
+
+    const currentTime = dayjs().unix()
+    const secondsFromNow = expectedFillTimeSec - currentTime
+
+    return dayjs.duration(secondsFromNow, 'seconds').humanize(true)
+  }, [expectedFillTimeSec])
+
+  return estimatedTimeDisplay
+}