+++ apps/web/src/views/SwapSimplify/hooks/useIsPriceImpactTooHigh.ts
@@ -3,21 +3,29 @@ import { useMemo, useRef } from 'react'
 import { warningSeverity } from 'utils/exchange'
 import { InterfaceOrder, isBridgeOrder, isXOrder } from 'views/Swap/utils'
 
-import { computeTradePriceBreakdown } from '../../Swap/V3Swap/utils/exchange'
+import { computeBridgeOrderFee } from 'views/Swap/Bridge/utils'
+import { computeTradePriceBreakdown, findHighestPriceImpact } from '../../Swap/V3Swap/utils/exchange'
 
 export const useIsPriceImpactTooHigh = (bestOrder: InterfaceOrder | undefined, isLoading?: boolean) => {
   const { chainId } = useActiveChainId()
   const chainIdRef = useRef(chainId)
 
-  const { priceImpactWithoutFee } = useMemo(
+  const priceBreakdown = useMemo(
     () =>
-      // TODO: remove when bridge is implemented
       isBridgeOrder(bestOrder)
-        ? { priceImpactWithoutFee: null }
+        ? computeBridgeOrderFee(bestOrder)
         : computeTradePriceBreakdown(isXOrder(bestOrder) ? bestOrder?.ammTrade : bestOrder?.trade),
     [bestOrder],
   )
-  const isPriceImpactTooHigh = useMemo(() => {
+  return useMemo(() => {
+    let priceImpactWithoutFee
+
+    if (Array.isArray(priceBreakdown)) {
+      priceImpactWithoutFee = findHighestPriceImpact(priceBreakdown)
+    } else {
+      priceImpactWithoutFee = priceBreakdown.priceImpactWithoutFee
+    }
+
     const warningLevel = warningSeverity(priceImpactWithoutFee)
     if (chainIdRef?.current === chainId) {
       if (!isLoading) return warningLevel >= 3
@@ -25,7 +33,5 @@ export const useIsPriceImpactTooHigh = (bestOrder: InterfaceOrder | undefined, i
     }
     chainIdRef.current = chainId
     return false
-  }, [priceImpactWithoutFee, chainId, isLoading])
-
-  return { isPriceImpactTooHigh }
+  }, [priceBreakdown, chainId, isLoading])
 }